Make Default LDAP attributes overridable by additional attributes
This request is designed to allow attributes specified in 'LDAP_ADDITIONAL_ATTRIBUTES' to override the default attributes.

For example, if you put 'uidnumber' in additional attributes, then instead of using LastUID+1 it will allow you to manually enter a UID number. 
This also allows you to re-label fields (eg, make the uid field into 'UNIX Username' by defining it as such in additional attributes)

It does not address the cn=lastUID being incremented, but logically an org that is manually assigning UIDs won't be relying on that anyway (Eg, they are using HR assigned badge-numbers as UNIX UIDs, so a manually entered UID will always be unique). 

Mechanisim is as follows:
1) Force both array's keys (but not values) to lowercase
2) Use array_replace_recursive instead of array_merge, to overwrite the 'default' array's key-value pairs with those of the additional_attributes array.